[batch] Split instance_pool into instance_pool and instance_monitor
